+++ src/app/programs/[slug]/page.tsx
@@ -1,3 +1,4 @@
 import ProgramBlogPage from './ProgramBlogPage';
+export { generateStaticParams, dynamicParams } from './ProgramBlogPage';
 
 export default ProgramBlogPage;